+++ lib/features/todo/domain/bloc/todo_list_bloc.dart
@@ -2,7 +2,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:todo_list/config/logging/logger.dart';
 import 'package:todo_list/core/services/shared_preferences_service.dart';
-import 'package:todo_list/features/todo/data/todo_repository.dart';
+import 'package:todo_list/features/todo/data/repository/todo_repository.dart';
 import 'package:todo_list/features/todo/domain/bloc/todo_list_event.dart';
 import 'package:todo_list/features/todo/domain/bloc/todo_list_state.dart';
 import 'package:todo_list/features/todo/domain/entity/todo.dart';
