--- lib/features/todo/presentation/todo_single/cubit/single_todo_state.dart
@@ -1,7 +1,0 @@
-import 'package:todo_list/features/todo/domain/entity/todo.dart';
-
-class SingleTodoState {
-  final Todo todo;
-
-  const SingleTodoState({required this.todo});
-}
